chore: renamed files to follow ts naming conventions
