+++ src/controllers/auth.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
-import { PasswordResetService } from '../services/passwordResetService';
-import { ForgotPasswordRequest, ResetPasswordRequest, ApiResponse } from '../types/auth';
+import { PasswordResetService } from '../services/password-reset.service';
+import { ForgotPasswordRequest, ResetPasswordRequest, ApiResponse } from '../types/auth.type';
 
 export class AuthController {
   private passwordResetService: PasswordResetService;
